Move OAuthRouter, OAuthConfig, TokenConfig etc to separate files
